# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Section {
  id: number;
  name: string;
  /** @nullable */
  nameAr?: string | null;
  slug: string;
  icon: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  order: number;
  isActive: boolean;
  createdAt: string;
}

export interface SectionInput {
  name: string;
  nameAr?: string;
  slug: string;
  icon: string;
  description?: string;
  descriptionAr?: string;
  order?: number;
  isActive?: boolean;
}

export interface SectionUpdate {
  name?: string;
  nameAr?: string;
  slug?: string;
  icon?: string;
  description?: string;
  descriptionAr?: string;
  order?: number;
  isActive?: boolean;
}

export interface Platform {
  id: number;
  name: string;
  /** @nullable */
  nameAr?: string | null;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** telegram_bot | website | program | tool */
  type: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sectionId: number;
  isActive: boolean;
  isFeatured: boolean;
  order?: number;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface PlatformInput {
  name: string;
  nameAr?: string;
  slug: string;
  description?: string;
  descriptionAr?: string;
  type: string;
  url?: string;
  imageUrl?: string;
  sectionId: number;
  isActive?: boolean;
  isFeatured?: boolean;
  order?: number;
  tags?: string;
}

export interface PlatformUpdate {
  name?: string;
  nameAr?: string;
  slug?: string;
  description?: string;
  descriptionAr?: string;
  type?: string;
  url?: string;
  imageUrl?: string;
  sectionId?: number;
  isActive?: boolean;
  isFeatured?: boolean;
  order?: number;
  tags?: string;
}

export interface NewsPost {
  id: number;
  title: string;
  /** @nullable */
  titleAr?: string | null;
  slug: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  contentAr?: string | null;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isPinned: boolean;
  isPublished: boolean;
  /** @nullable */
  category?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface NewsPostInput {
  title: string;
  titleAr?: string;
  slug: string;
  content?: string;
  contentAr?: string;
  excerpt?: string;
  imageUrl?: string;
  isPinned?: boolean;
  isPublished?: boolean;
  category?: string;
}

export interface NewsPostUpdate {
  title?: string;
  titleAr?: string;
  slug?: string;
  content?: string;
  contentAr?: string;
  excerpt?: string;
  imageUrl?: string;
  isPinned?: boolean;
  isPublished?: boolean;
  category?: string;
}

export interface Advertisement {
  id: number;
  title: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  position?: string | null;
  isActive: boolean;
  order?: number;
  createdAt: string;
}

export interface AdvertisementInput {
  title: string;
  imageUrl?: string;
  linkUrl?: string;
  position?: string;
  isActive?: boolean;
  order?: number;
}

export interface AdvertisementUpdate {
  title?: string;
  imageUrl?: string;
  linkUrl?: string;
  position?: string;
  isActive?: boolean;
  order?: number;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  /** @nullable */
  siteNameAr?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  secondaryColor?: string | null;
  /** @nullable */
  accentColor?: string | null;
  /** @nullable */
  heroTitle?: string | null;
  /** @nullable */
  heroTitleAr?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  /** @nullable */
  heroSubtitleAr?: string | null;
  /** @nullable */
  sumerianText?: string | null;
  /** @nullable */
  footerText?: string | null;
  /** @nullable */
  footerTextAr?: string | null;
  /** @nullable */
  socialTelegram?: string | null;
  /** @nullable */
  socialTwitter?: string | null;
  /** @nullable */
  socialYoutube?: string | null;
  /** @nullable */
  adminPassword?: string | null;
  maintenanceMode?: boolean;
}

export interface SiteSettingsUpdate {
  siteName?: string;
  siteNameAr?: string;
  logoUrl?: string;
  faviconUrl?: string;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  heroTitle?: string;
  heroTitleAr?: string;
  heroSubtitle?: string;
  heroSubtitleAr?: string;
  sumerianText?: string;
  footerText?: string;
  footerTextAr?: string;
  socialTelegram?: string;
  socialTwitter?: string;
  socialYoutube?: string;
  adminPassword?: string;
  maintenanceMode?: boolean;
}

export type PlatformStatsPlatformsByType = {[key: string]: number};

export type PlatformStatsPlatformsBySectionItem = {
  sectionId: number;
  sectionName: string;
  count: number;
};

export interface PlatformStats {
  totalSections: number;
  totalPlatforms: number;
  totalNews: number;
  featuredPlatforms: number;
  activePlatforms: number;
  platformsByType?: PlatformStatsPlatformsByType;
  platformsBySection?: PlatformStatsPlatformsBySectionItem[];
}

export type ListPlatformsParams = {
sectionId?: number;
featured?: boolean;
};

export type ListNewsParams = {
pinned?: boolean;
limit?: number;
};

export type ListAdvertisementsParams = {
active?: boolean;
};