# lib/api-zod/src/generated/types/advertisement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Advertisement {
  id: number;
  title: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  position?: string | null;
  isActive: boolean;
  order?: number;
  createdAt: string;
}